+++ changelog.md
@@ -4,6 +4,7 @@
 * Code cleanup & improved type annotations in common.
 * Uses multiline textboxes & labels if possible.
 * Serializes blenders "Empty" display settings.
+* `stfexp.instance.text` bugfix.
 
 ## v0.1.10
 * Poll function bugfixes when checking whether a bone is selected.